requirements.txt

fastapi
uvicorn
langchain
langchain-google-genai
langsmith
tavily-python
langgraph
pydantic
python-dotenv
# pandas
# openpyxl
# xlrd
langchain-community
langchain-core
python-multipart
faiss-cpu
requests
huggingface_hub



# pip install fastapi uvicorn langchain langchain-community langchain-google-genai python-multipart dotenv
# pip install python-docx
# pip install python-pptx
# pip install PyPDF2
# pip install openpyxl
# pip install python-docx PyPDF2 python-pptx openpyxl


# fastapi
# uvicorn
# langchain
# langchain-community
# langchain-google-genai
# python-dotenv
# pandas
# python-docx
# PyPDF2
# python-pptx
# openpyxl

#  pip install fastapi uvicorn langchain langsmith tavily-python google-generativeai

# pip install faiss-cpu langchai
# pip install huggingface_hub



HUGGINGFACE_API_TOKEN=YOUR_HF_TOKEN







